Always provide a user context to signal handlers
Signal handlers expect some context.  They actually expect a struct
ucontext, which is defined in Glibc.  Further, some programs make
assumptions about the contents of ucontext (and mcontext).  We'll provide
them with a user_context for now.  It's one thing to want a ucontext.  It's
a bit nastier for handlers to demand that it matches the format of whatever
glibc they are using (where the individual fields of an x86_64 mcontext are
API).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>