Fix sbrk's lock initialization (XCC)
We were getting away with the uninitialized lock since a value of 0 was OK
for the current LLL locks.

I spotted this when trying to use PDR locks, and LD was flipping out on a
bare-bones replacement for spin_pdr_locks.  A while loop and a CTRL-B
backtrace idenfitied sbrk() as the culprit.

Rebuild glibc, if you want.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>