Add a printx lock
When debugging and relying on printed output, if you have a process that is
spamming the console, it can be hard to see what the kernel is printing
out.

This lock is meant to be used in debugging.  If you have printx turned on,
(px from the monitor), then the console prints will be synchronized with
other uses of px_lock().

For instance, if you want to drop a backtrace at some point, you could do:

sys_foo():
	if (some_condition && printx_on) {
		px_lock();
		backtrace_user_ctx(current, current_ctx);
		px_unlock();
	}

You'd never commit that blob of code, but it's useful when tracking down a
bug.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>