qio: Fix copyblock()
Previously, copyblock() couldn't handle extra data.

The only caller was passing the full block length, so let's just always do
the full amount.  At this point, it's basically the same as
linearizeblock(), minus the block swapping.  So we can implement
linearizeblock() with copyblock().

Note that Qsnoop uses copyblock(), making a copy of the block to pass up
to a snooping process (e.g. snoopy).  I considered making something like
qclone, where we clone the block instead - using pointers to the original
block's memory.  This would probably work with our current setup, but if we
ever use user memory for the block extra data segments, then we'd have a
problem.  The issue is one of immutability of data, and if we are using
user memory, then we have no guarantee of immutability (unless it's
read-only or something).  For now, we can just make a copy.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>