Use a "one block at a time" policy for snoop queues
This fixes snoopy's packet loss.

Snoopy expects to receive a packet at a time.  If you send more than one
packet in a read, the later packets get dropped.  We could change snoopy,
but it is a little easier to tell the kernel to just return one block
(packet) at a time when we do a qread.

In part, I went with this approach since we'll probably want more kernel
support for packet tracing, not less.  For instance, right now, snoopy's
time stamps are taken in userspace when the queue is read, not when the
packet is received or sent.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>