Change syscall usec timeouts to unsigned longs
I noticed this due to some sys_block calls having a 'negative' argument in
the printout (due to the %d in the saved string).

While I was here, I also changed halt_core, though note that that timeout
was more of a 'future plan', I think.  The code doesn't use it.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>