Delay clearing owning proc in sys_exec
If we do it before any of the return calls, we could end up returning to
userspace while owning_proc isn't set.  I think the rest of the kernel is
able to handle this, but there's no sense messing around.  The old comment
makes it sound like we can block in that state too, which is probably true,
but returning by anything other than the error path ways seems like a bad
idea.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>