mlx4: Remove page_is_free() safety check
With the arena allocator, it won't be easy to query the state of a given
page.  We actually could do that, if we wanted, with an arena helper
that looks up the btag for a given address, but it's a pain, it won't be
fast, and it will probably not work well with NUMA.

Considering this style of page pinning needs to change anyways, we might
as well remove it.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>