vmm: allow a vmm to override the vmcall function
Add a vmcall struct to the guest thread struct. This
allows us, on a guest thread by guest thread basis, to
support vmcalls.

I've tested this with dune and it works fine.
Longer term, we may want to define an ops structure
but I think that's rushing it a bit.

Change-Id: Ic381f0e70946ba2396303e5d6428bc999ec4b6dd
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>